Add check for currentUser review

+++ MovieApp/Review.cs
@@ -4,10 +4,10 @@ namespace MovieApp
     internal struct Review
     {
         public string Username { get; set; }
-        public double Rating { get; set; }
+        public int Rating { get; set; }
         public string Comment { get; set; }
 
-        public Review ( string username, double rating, string comment ) {
+        public Review ( string username, int rating, string comment ) {
             Username = username;
             Rating = rating;
             Comment = comment;
